doc: recommend treating CS9057 as error
CS9057 is a compiler warning issued when a consuming project is using
an earlier version of the .NET compiler than was used to compile the
Introspection generator. Such a mismatch can have serious,
hard-to-diagnose downstream consequences for users. This change
recommends treating CS9057 as an error in consuming projects, to make
those issues easier to diagnose and fix.

Additionally, this change updates a VSCode setting to reflect new
naming.